Fold source-repo capture test into e2e_test.go
